refactor auth context helpers
