+++ README.md
@@ -141,6 +141,8 @@ Create these keys by signing up for a developer account at [Unsplash](https://un
 The default credentials are **demo@example.com** / **password**. You can also
 register a new account using the **Sign Up** page. Successful authentication
 returns a token stored in `localStorage` by the `AuthProvider`.
+Access this context via the `useAuth` hook exported from
+`src/contexts/authHelpers.js`.
 
 During development the API server runs on `http://localhost:3001`, so the
 `VITE_API_BASE_URL` variable should point there. For production builds set it to
