+++ src/screens/Login.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAuth } from '../contexts/AuthProvider';
+import { useAuth } from '../contexts/authHelpers';
 
 export default function Login() {
   const { login } = useAuth();
