+++ src/screens/SignUp.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate, Link } from 'react-router-dom';
-import { useAuth } from '../contexts/AuthProvider';
+import { useAuth } from '../contexts/authHelpers';
 
 export default function SignUp() {
   const { register } = useAuth();
